==> 1_Data_Processing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import networkx as nx
%matplotlib inline

file_path = './Data/letterbox_anonym.csv'
lboxd_data = pd.read_csv(file_path, sep=';', index_col=0)

lboxd_data

#### Looking at the dataset
- We have ratings from 0 to 5, 0 probably meaning that the movie was watched but not rated by the user, still useful for community detection though
- We have 108'276 movies
- We have 557 users

print('Unique ratings:', sorted(lboxd_data['rating'].unique()))
print('Number of unique movies:', len(lboxd_data['title'].unique()))
print('Number of unique users:', len(lboxd_data['user'].unique()))

There are 22 missing movies
- The rows are only missing the title of the movie, the user id and rating is present.
- Not sure what caused this problem but it is not possible to infer the titles as users are anonymous and the dataset doesn't have supporting data.

for column in lboxd_data.columns:

    i_nans = lboxd_data[lboxd_data[column].isna()].index.tolist()
    n_nans = sum(lboxd_data[column].isna())
    print(f'[{column}] column has [{n_nans}] missing values.')
    print(lboxd_data.iloc[i_nans, :])

Removing the rows with missing values
- clean_lboxd has 0 missing values in all columns.
- clean_lboxd has 22 less rows than the original dataset.

clean_lboxd = lboxd_data.dropna()

for column in clean_lboxd.columns:

    n_nans = sum(clean_lboxd[column].isna())
    print(f'[{column}] column has [{n_nans}] missing values.')

print('Number of ratings lost:',len(lboxd_data) - len(clean_lboxd))

Casting the columns to their appropriate datatypes
- To ensure that we have no problems with the types

clean_lboxd = clean_lboxd.astype({'user':int, 'title':str, 'rating':float})
clean_lboxd.dtypes

clean_lboxd

#### Adding an ID columns as an alternative
- In case we feel that some movie titles are hard to work with
- Starting at 1000 to ensure there is no overlapping with the user ids (go up to 557)

# Add a unique ID for each movie starting at 1000 to ensure that the user IDs are not the same as the movie IDs
# Each movie title has to be assigned an unique ID
clean_lboxd['movie_id'] = clean_lboxd['title'].astype('category').cat.codes + 1000
clean_lboxd

# Check the minimum and maximum movie IDs
print(clean_lboxd['movie_id'].min())
print(clean_lboxd['movie_id'].max())

# Check minimum and maximum user IDs
print(clean_lboxd['user'].min())
print(clean_lboxd['user'].max())

# Check the number of unique movie IDs
len(clean_lboxd['movie_id'].unique())

# Check which movie title corresponds to the movie ID 1000
print(clean_lboxd[clean_lboxd['movie_id'] == 1000])

# Check how many times that movie_id appears in the dataset
print(clean_lboxd['movie_id'].value_counts()[1000])

#### Sampling from the dataset
- Our dataset is quite large (1M4 ratings), and while testing bilouvain we found that runtime was extremely long (didn't converge after almost 20h)
- To ensure that we can test our algorithm we need to sample the dataset, however it's hard to do sampling while keeping network structure.
- A way I thought of doing so is by removing movies that appear only once in the dataset (meaning only one user rated them) as they might not have a significant impact on the communities.

# Make a new df without the movies that appear only once
movies_count = clean_lboxd['title'].value_counts()
movies_to_keep = movies_count[movies_count > 1].index
new_lboxd_data = clean_lboxd[clean_lboxd['title'].isin(movies_to_keep)]
new_lboxd_data

len(clean_lboxd) - len(new_lboxd_data)

# Export the new df to a csv file
new_lboxd_data.to_csv('./Data/lboxd_no_single_IDcoded.csv', index=False)

# Export the original clean df to a csv file
clean_lboxd.to_csv('./Data/lboxd_IDcoded.csv', index=False)

# Convert the new_lboxd_data to a networkx bipartite graph with users and movies as nodes
G = nx.Graph()
G.add_nodes_from(new_lboxd_data['user'].unique(), bipartite=0)
G.add_nodes_from(new_lboxd_data['title'].unique(), bipartite=1)
G.add_edges_from(new_lboxd_data[['user', 'title']].values)

# Check if the graph is bipartite
nx.is_bipartite(G)

# Check the number of nodes and edges
print(nx.number_of_nodes(G))
print(nx.number_of_edges(G))

# Get the sets of nodes
users = {n for n, d in G.nodes(data=True) if d['bipartite'] == 0}
movies = {n for n, d in G.nodes(data=True) if d['bipartite'] == 1}

# Print the number of nodes in the sets
print(len(users))
print(len(movies))
