==> 2_Data_Augmentation.ipynb <==
import numpy as np
import pandas as pd
import networkx as nx
import matplotlib.pyplot as plt
import time
import os

data_filepath = "./Data/letterbox_anonym.csv"
letterboxd = pd.read_csv(data_filepath, sep=";", index_col=0)

letterboxd

clean_lbxd = letterboxd.dropna()
clean_lbxd = clean_lbxd.rename(columns={'title': 'name'})
clean_lbxd

len(clean_lbxd['name'].unique())

for rating in sorted(clean_lbxd['rating'].unique()):
    
    n_ratings = len(clean_lbxd[clean_lbxd['rating'] == rating])
    print(f"{rating}* -> {n_ratings}")

### Movie Databases

# Ended up not needing to use these databases (as they don't contain that many movies we have)

#mov_imdb_path = "./Dataset/movies_imdb.csv"
#mov_imdb = pd.read_csv(mov_imdb_path)

#mov_data_path = "./Dataset/movie_data.csv"
#mov_data = pd.read_csv(mov_data_path)

**Lbxd Movie Data**

Datasets not added to github because way too big !

mov_lbxd_path = "./Data/movies.csv"
mov_lbxd = pd.read_csv(mov_lbxd_path)

genre_lbxd_path = "./Data/lbxd_genres.csv"
genre_lbxd = pd.read_csv(genre_lbxd_path)

countries_lbxd_path = "./Data/lbxd_countries.csv"
countries_lbxd = pd.read_csv(countries_lbxd_path)

mov_lbxd = mov_lbxd.drop(['tagline', 'description'], axis=1)
mov_lbxd

genre_lbxd

countries_lbxd

**Augmenting the dataset of lbxd movies with genre and country**

# Merging the movies from letterboxd dataset with the genres letterboxd dataset
merged_df = pd.merge(mov_lbxd, genre_lbxd, on='id', how='outer')

# Grouping according to id (because there are duplicate ids on genre)
# then aggregating according the name, date, minute, rating, and genre as a list to contain all genres
merged_df = merged_df.groupby(['id'], as_index=False).agg({
    'name': 'first',
    'date': 'first',
    'minute': 'first',
    'rating': 'first',
    'genre': list
})
merged_df

# Same as above but by using the previously merged dataframe to add the countries
augm_mov_lbxd = pd.merge(merged_df, countries_lbxd, on='id', how='outer')
augm_mov_lbxd = augm_mov_lbxd.groupby(['id'], as_index=False).agg({
    'name': 'first',
    'date': 'first',
    'minute': 'first',
    'rating': 'first',
    'genre': 'first', # can't set it as a list again, else it duplicates the values
    'country': list
})

# This final dataset contains movie names, their genre (if it was found) and their country (if it was found)
# Sadly we have missing values for a lot of the information -> date, minute, rating, genre, country
augm_mov_lbxd = augm_mov_lbxd.rename(columns={'rating': 'mean_rating'})
augm_mov_lbxd

### Checking how many movies are inside the lbxd movie database

# Unique movie names from letterboxd ratings dataset
lbxd_movies = clean_lbxd['name'].unique()

start = time.time()
movies_fromdb = augm_mov_lbxd['name'].unique()
movies_in_db = []
movies_not_db = []
lbxd_movies = clean_lbxd['name'].unique()
moviesdb_set = set(movies_fromdb) # Improves performance from 17min to 0.15 seconds
for movie in lbxd_movies:
    if movie in moviesdb_set:
        movies_in_db.append(movie)
    else:
        movies_not_db.append(movie)

print(len(movies_in_db))
print(len(movies_not_db))
end = time.time()
total_time = end - start
print(f"Total time: {total_time/60}")

### Merging our ratings dataset with the augmented lbxd movie database according to the movie name
- 3595 movies could not be found on the movie database
- From the 104680 that were found, some can be missing values in ['date', 'minute', 'mean_rating', 'genre', 'country']
- In conclusion we have a lot of missing values, however this can still be useful for visualizing the communities later on and there's not much I can do to compensate for these missing values (as some movies don't have that inforamtion on databases and doing it manually would take too much time)

movies_in_db_df = augm_mov_lbxd[augm_mov_lbxd['name'].isin(lbxd_movies)]

augm_df = pd.merge(clean_lbxd['name'], movies_in_db_df[['name', 'date', 'minute', 'mean_rating', 'genre', 'country']], on='name', how='left')

augm_df = augm_df.groupby(['name'], as_index=False).agg({
    'date': 'first',
    'minute': 'first',
    'mean_rating': 'first',
    'genre': 'first',
    'country': 'first',

})

augm_df = augm_df.replace({None: np.nan})
augm_df

# Printing the number of missing values for each column

for column in augm_df.columns:

    miss_vals = sum(augm_df[column].isna())

    print(f"{column}: {miss_vals} missing values")

### Saving augmented movies dataframe to csv

# Saving the augmented movies dataframe to a csv file
augm_df.to_csv('./Data/augmented_movies.csv')
