==> 3_Data_Augmentation2.ipynb <==
import pandas as pd
import numpy as np
import networkx as nx
import matplotlib.pyplot as plt
import seaborn as sns

augm_path = './Data/augmented_movies.csv'
lboxd_path = './Data/lboxd_IDcoded.csv'

augm_df = pd.read_csv(augm_path, index_col=0)
lboxd_df = pd.read_csv(lboxd_path)

augm_df.rename(columns={'name':'title'}, inplace=True)
augm_df

lboxd_df

augm_merged = pd.merge(lboxd_df, augm_df, on='title')
augm_merged

augm_merged.to_csv('./Data/augm_lboxd.csv', index=False)
